Queue is commented out to await more time for implementation; convert tree to use uint instead of uint64

+++ tree/index_test.go
@@ -13,7 +13,7 @@ func TestFind(t *testing.T) {
 
 	tests := map[string]struct {
 		index   index
-		argID   uint64
+		argID   uint
 		expNode Node
 	}{
 		"nil index": {
